Move the logic for scanning SSE to the common package
To use it from engine as well.
