+++ common/pkg/sse/const.go
@@ -1,0 +1,6 @@
+package sse
+
+const (
+	// DoubleNewline is a double newline.
+	DoubleNewline = "\n\n"
+)
