+++ server/internal/server/completions.go
@@ -1,8 +1,6 @@
 package server
 
 import (
-	"bufio"
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -11,6 +9,7 @@ import (
 
 	"github.com/llm-operator/common/pkg/id"
 	v1 "github.com/llm-operator/inference-manager/api/v1"
+	"github.com/llm-operator/inference-manager/common/pkg/sse"
 	"github.com/llm-operator/inference-manager/server/internal/infprocessor"
 	mv1 "github.com/llm-operator/model-manager/api/v1"
 	"github.com/llm-operator/rbac-manager/pkg/auth"
@@ -126,12 +125,10 @@ func (s *S) CreateChatCompletion(
 		return
 	}
 
-	scanner := bufio.NewScanner(resp.Body)
-	scanner.Buffer(make([]byte, 4096), 4096)
-	scanner.Split(split)
+	scanner := sse.NewScanner(resp.Body)
 	for scanner.Scan() {
 		b := scanner.Text()
-		if _, err := w.Write([]byte(b + "\n\n")); err != nil {
+		if _, err := w.Write([]byte(b + sse.DoubleNewline)); err != nil {
 			http.Error(w, err.Error(), http.StatusInternalServerError)
 			return
 		}
@@ -142,40 +139,4 @@ func (s *S) CreateChatCompletion(
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
-}
-
-// split tokenizes the input. The arguments are an initial substring of the remaining unprocessed data and a flag,
-// atEOF, that reports whether the Reader has no more data to give.
-// The return values are the number of bytes to advance the input and the next token to return to the user,
-// if any, plus an error, if any.
-func split(data []byte, atEOF bool) (int, []byte, error) {
-	if atEOF {
-		var nextToken []byte
-		if len(data) > 0 {
-			nextToken = data
-		}
-		return len(data), nextToken, nil
-	}
-
-	// Find a double newline.
-	delims := [][]byte{
-		[]byte("\r\r"),
-		[]byte("\n\n"),
-		[]byte("\r\n\r\n"),
-	}
-	pos := -1
-	var dlen int
-	for _, d := range delims {
-		n := bytes.Index(data, d)
-		if pos < 0 || (n >= 0 && n < pos) {
-			pos = n
-			dlen = len(d)
-		}
-	}
-
-	if pos >= 0 {
-		return pos + dlen, data[0:pos], nil
-	}
-
-	return 0, nil, nil
 }
