+++ server/internal/server/completions_test.go
@@ -1,7 +1,6 @@
 package server
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"encoding/json"
@@ -20,28 +19,6 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
-
-func TestSplit(t *testing.T) {
-	input := `line0
-
-line1
-
-line2
-
-line3`
-	scanner := bufio.NewScanner(bytes.NewReader([]byte(input)))
-	scanner.Buffer(make([]byte, 4096), 4096)
-	scanner.Split(split)
-	var got []string
-	for scanner.Scan() {
-		got = append(got, scanner.Text())
-	}
-	err := scanner.Err()
-	assert.NoError(t, err)
-
-	want := []string{"line0", "line1", "line2", "line3"}
-	assert.ElementsMatch(t, want, got)
-}
 
 func TestCreateChatCompletion(t *testing.T) {
 	const (
